refactor(chat): split slash command modules
