+++ src/chat-commands-contract.ts
@@ -1,0 +1,44 @@
+import type { ChatRow } from "./chat-contract";
+import type { Client } from "./client-contract";
+import type { ConfigScope } from "./config-contract";
+import type { Session, SessionState, SessionTokenUsageEntry } from "./session-contract";
+
+export type CommandResult = {
+  stop: boolean;
+  userText: string;
+};
+
+export type CommandContext = {
+  text: string;
+  resolvedText: string;
+  client: Client;
+  store: SessionState;
+  currentSession: Session;
+  setCurrentSession: (next: Session) => void;
+  setTokenUsage?: (updater: (current: SessionTokenUsageEntry[]) => SessionTokenUsageEntry[]) => void;
+  toRows: (messages: Session["messages"]) => ChatRow[];
+  setRows: (updater: (current: ChatRow[]) => ChatRow[]) => void;
+  setShowHelp: (updater: (current: boolean) => boolean) => void;
+  setValue: (next: string) => void;
+  persist: () => Promise<void>;
+  exit: () => void;
+  openSkillsPanel: () => Promise<void>;
+  openResumePanel: () => void;
+  openModelPanel: () => void | Promise<void>;
+  persistModelConfig?: (key: string, value: string, scope: ConfigScope) => Promise<void>;
+  activateSkill?: (skillName: string, args: string) => Promise<boolean>;
+  startAssistantTurn?: (userText: string) => Promise<void>;
+  clearTranscript: (sessionId?: string) => void;
+  tokenUsage: SessionTokenUsageEntry[];
+  memoryApi?: {
+    listMemories: typeof import("./memory-ops").listMemories;
+    addMemory: typeof import("./memory-ops").addMemory;
+    removeMemory: typeof import("./memory-ops").removeMemory;
+  };
+};
+
+export type SlashCommand = {
+  name: string;
+  match: (value: string) => boolean;
+  run: () => Promise<CommandResult>;
+};
